@@ -9,16 +9,22 @@ use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\URL;
 
 // Global Helpers
-if ( !function_exists('locales') ) { function locales() { return config('app.available_locales'); } }
+if (!function_exists('locales')) {
+    function locales()
+    {
+        return config('app.available_locales');
+    }
+}
 
 // Panel Helpers
-if(! function_exists('isActive') ) {
-    function isActive($key , $activeClassName = 'active', $default = '') {
-        if(is_array($key))
-            return in_array(Route::currentRouteName() , $key) ? $activeClassName : $default;
+if (!function_exists('isActive')) {
+    function isActive($key, $activeClassName = 'active', $default = '')
+    {
+        if (is_array($key))
+            return in_array(Route::currentRouteName(), $key) ? $activeClassName : $default;
 
         // check for naming patterns like: module.moduleName.*
-        if(!is_array($key) && strpos($key, '*') !== false)
+        if (!is_array($key) && strpos($key, '*') !== false)
             return (strpos(Route::currentRouteName(), str_replace('*', '', $key)) !== false) ? $activeClassName :  $default;
 
         return Route::currentRouteName() == $key ? $activeClassName :  $default;
@@ -26,8 +32,9 @@ if(! function_exists('isActive') ) {
 }
 
 // Route helpers
-if(! function_exists('isCurrentUrl') ) {
-    function isCurrentUrl(string $current_url, string $active_class_name = 'active', string $default = '') {
+if (!function_exists('isCurrentUrl')) {
+    function isCurrentUrl(string $current_url, string $active_class_name = 'active', string $default = '')
+    {
         return URL::current() == $current_url ? $active_class_name : $default;
     }
 }
@@ -35,70 +42,81 @@ if(! function_exists('isCurrentUrl') ) {
 // Panel-Theme helpers
 
 if (!function_exists('add_panel_menu_item')) {
-    function add_panel_menu_item(string $menu_item_route, string $menu_item_icon, string $menu_item_text, string|null $menu_item_gate = null, int $priority = 10, array $menu_item_extra = []) {
-        Hooks::addAction('panel.menu_items', function() use($menu_item_gate, $menu_item_route, $menu_item_icon, $menu_item_text, $menu_item_extra) {
+    function add_panel_menu_item(string $menu_item_route, string $menu_item_icon, string $menu_item_text, string|null $menu_item_gate = null, int $priority = 10, array $menu_item_extra = [])
+    {
+        Hooks::addAction('panel.menu_items', function () use ($menu_item_gate, $menu_item_route, $menu_item_icon, $menu_item_text, $menu_item_extra) {
             echo panelView('menu-item', [
-                'menu_item_gate'=> $menu_item_gate,
-                'menu_item_route'=> $menu_item_route,
-                'menu_item_icon'=> $menu_item_icon,
-                'menu_item_text'=> $menu_item_text,
-                'menu_item_extra'=> $menu_item_extra
+                'menu_item_gate' => $menu_item_gate,
+                'menu_item_route' => $menu_item_route,
+                'menu_item_icon' => $menu_item_icon,
+                'menu_item_text' => $menu_item_text,
+                'menu_item_extra' => $menu_item_extra
             ]);
         }, $priority);
     }
 }
 
 if (!function_exists('viewIfExist')) {
-    function viewIfExist($view_name, $variables = [], $error_view = 'errors.theme404') {
-        if ( view()->exists($view_name) ) {
+    function viewIfExist($view_name, $variables = [], $error_view = 'errors.theme404')
+    {
+        if (view()->exists($view_name)) {
             return view($view_name, $variables);
         }
         return view($error_view, compact('view_name'));
     }
 }
 if (!function_exists('themeView')) {
-    function themeView($view, $variables = []) {
-        return viewIfExist('themes.'.Theme::currentName().'.'.$view, $variables);
+    function themeView($view, $variables = [])
+    {
+        return viewIfExist('themes.' . Theme::currentName() . '.' . $view, $variables);
     }
 }
 if (!function_exists('themeAsset')) {
-    function themeAsset($path, $secure = null) {
-        return app('url')->asset('themes/'.Theme::currentName().'/'.$path, $secure);
+    function themeAsset($path, $secure = null)
+    {
+        return app('url')->asset('themes/' . Theme::currentName() . '/' . $path, $secure);
     }
 }
 
 if (!function_exists('panelLayout')) {
-    function panelLayout() {
-        return 'panel.layouts.'.Theme::panelTheme().'.layout';
+    function panelLayout()
+    {
+        return 'panel.layouts.' . Theme::panelTheme() . '.layout';
     }
 }
 
 if (!function_exists('panelView')) {
-    function panelView($view, $variables = []) {
-        return viewIfExist('panel.layouts.'.Theme::panelTheme().'.'.$view, $variables);
+    function panelView($view, $variables = [])
+    {
+        return viewIfExist('panel.layouts.' . Theme::panelTheme() . '.' . $view, $variables);
     }
 }
 if (!function_exists('panelAsset')) {
-    function panelAsset($path, $secure = null) {
-        return app('url')->asset('themes/'.Theme::panelTheme().'/'.$path, $secure);
+    function panelAsset($path, $secure = null)
+    {
+        return app('url')->asset('themes/' . Theme::panelTheme() . '/' . $path, $secure);
     }
 }
 
 if (!function_exists('themeTemplates')) {
-    function themeTemplates() {
-        $path = '../resources/views/themes/'.Theme::currentName().'/templates';
+    function themeTemplates()
+    {
+        $path = '../resources/views/themes/' . Theme::currentName() . '/templates';
 
-        try { $files = scandir($path); }
-        catch (\Throwable $th) { return null; }
+        try {
+            $files = scandir($path);
+        } catch (\Throwable $th) {
+            return null;
+        }
 
         $templates = [];
 
         foreach ($files as $key => $file) {
-            if (is_dir($path.'/'.$file)) continue;
+            if (is_dir($path . '/' . $file)) continue;
 
-            $firstLine = trim(fgets(fopen($path.'/'.$file, 'r')));
+            $firstLine = trim(fgets(fopen($path . '/' . $file, 'r')));
 
-            $templates[str_replace('.blade.php', '',$file)] = grabThemeKey('name', $firstLine);
+            $templates[str_replace('.blade.php', '', $file)] = grabThemeKey('name', $firstLine);
         }
 
         return ($templates);
@@ -106,38 +124,41 @@ if (!function_exists('themeTemplates')) {
 }
 
 if (!function_exists('grabThemeKey')) {
-    function grabThemeKey($key, $string) {
+    function grabThemeKey($key, $string)
+    {
         preg_match('/{{--[\s\S](.*):[\s\S](.*)[\s\S]--}}/', $string, $output_array);
-        
+
         return count($output_array) == 3 && $output_array[1] == $key ? $output_array[2] : null;
     }
 }
 
 // Panel-User helpers
 if (!function_exists('checkGate')) {
-    function checkGate($gate) {
+    function checkGate($gate)
+    {
         if (gettype($gate) == 'array') {
-            foreach($gate as $value) {
+            foreach ($gate as $value) {
                 if (Gate::allows($value)) return true;
             }
             return false;
         }
 
         if (gettype($gate) == 'string') return Gate::allows($gate);
-        
+
         echo 'error@Helpers::checkGate : Wrong gate type';
         return false;
     }
 }
 
 // Option helpers
-if (! function_exists('get_option')) {
-    function get_option($tag, $default = null) {
+if (!function_exists('get_option')) {
+    function get_option($tag, $default = null)
+    {
         try {
-            return Cache::remember('Option::'.$tag.'-'.app()->getLocale(), 14400, function () use ($tag, $default) {
+            return Cache::remember('Option::' . $tag . '-' . app()->getLocale(), 14400, function () use ($tag, $default) {
                 $option = Option::whereTag($tag)->first();
                 if (is_null($option)) return $default;
-                
+
                 return ($option->is_translatable ? $option->data : $option->plain_data);
             });
         } catch (\Throwable $th) {
@@ -146,47 +167,50 @@ if (! function_exists('get_option')) {
     }
 }
 
-if (! function_exists('update_option')) {
-    function update_option($option, $data) {
+if (!function_exists('update_option')) {
+    function update_option($option, $data)
+    {
         if (gettype($option) == 'integer')
             $option = Option::find($option);
         elseif (gettype($option) == 'string')
             $option = Option::whereTag($option)->first();
-        
+
         if (is_null($option) || get_class($option) != 'App\Models\Option') {
             // if $option variable was string, it means it is actually holding the option tag
             if (gettype($option) == 'string') {
-                Option::create(array_merge($data, ['tag'=> $option]));
+                Option::create(array_merge($data, ['tag' => $option]));
             }
             return false;
         }
 
-        Cache::forget('Option::'.$option->tag.'-'.app()->getLocale());
+        Cache::forget('Option::' . $option->tag . '-' . app()->getLocale());
         $option->update($data);
         return true;
     }
 }
 
-if (! function_exists("santizieString") ) {
-    function santizieString($str) {
+if (!function_exists("santizieString")) {
+    function santizieString($str)
+    {
         $letters = [
             '–', '—', '"', '"', '"', '\'', '\'', '\'',
             '«', '»', '&', '÷', '>',    '<', '$', '/'
         ];
-    
+
         $str = str_replace($letters, " ", $str);
         $str = str_replace("&", "and", $str);
         $str = str_replace("?", "", $str);
         $str = strtolower(str_replace(" ", "-", $str));
-    
+
         return ($str);
     }
 }
 
-if (! function_exists("generateSlug") ) {
-    function generateSlug($value, $model = null, $current = null, $translatable = true, $column = 'slug') {
+if (!function_exists("generateSlug")) {
+    function generateSlug($value, $model = null, $current = null, $translatable = true, $column = 'slug')
+    {
         $value = santizieString($value == '' ? time() : $value);
-        
+
         if (!is_null($model)) {
             $model = (new $model)->query();
             if ($translatable) $model->whereTranslation($column, $value);
@@ -194,7 +218,7 @@ if (! function_exists("generateSlug") ) {
 
             if (!is_null($current)) $model->where('id', '!=', $current);
 
-            if ( $model->count() > 0 ) $value .= '-'.time();
+            if ($model->count() > 0) $value .= '-' . time();
         }
 
         return $value;
