@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Panel;
 
 use App\Http\Controllers\Controller;
+use Artesaos\SEOTools\Facades\SEOTools;
 use Illuminate\Http\Request;
 
 class PanelController extends Controller
@@ -13,10 +14,8 @@ class PanelController extends Controller
      */
     public function index()
     {
-        if (class_exists('\SEO')) \SEO::setTitle(__('Dashboard'));
-
-        // if (auth()->user()->role_id == 0) return view('panel::client-dashboard');
-
+        if (class_exists('\SEO')) SEOTools::setTitle(__('Dashboard'));
+        
         return view('panel.index');
     }
 }
