@@ -2,8 +2,8 @@
 
 namespace App\Providers;
 
+use App\Models\Permission;
 use App\Models\User;
-use App\Models\User\Permission;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Gate;
@@ -12,9 +12,9 @@ use Illuminate\Support\Facades\Schema;
 class AuthServiceProvider extends ServiceProvider
 {
     /**
-     * The policy mappings for the application.
+     * The model to policy mappings for the application.
      *
-     * @var array
+     * @var array<class-string, class-string>
      */
     protected $policies = [
         // 'App\Models\Model' => 'App\Policies\ModelPolicy',
@@ -22,39 +22,42 @@ class AuthServiceProvider extends ServiceProvider
 
     /**
      * Register any authentication / authorization services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
         $this->registerPolicies();
 
-        $this->app->scoped('userPermissions', function() {
-            $user = auth('sanctum')->user() ?? auth('web')->user() ?? null;
+        $this->app->scoped('userPermissions', function () {
+            $user = $this->getUser();
 
             if (is_null($user)) return [];
 
             // if user has no role , return empty array
             return $user->role()
-                    ->with('permissions')
-                    ->first()
-                    ?->permissions
-                    ->pluck('tag')
-                    ->toArray() ?? [];
+                ->with('permissions')
+                ->first()
+                ?->permissions
+                ->pluck('tag')
+                ->toArray() ?? [];
         });
 
         if (Schema::hasTable('permissions')) {
             $permissions = Cache::remember(
-                'allPermissions', 
-                14400, 
-                fn() => Permission::select(['id', 'tag'])->get()->pluck('tag')->toArray()
+                'allPermissions',
+                14400,
+                fn () => Permission::select(['id', 'tag'])->get()->pluck('tag')->toArray()
             );
 
-            foreach($permissions as $permission) {
+            foreach ($permissions as $permission) {
                 Gate::define($permission, function () use ($permission) {
                     return in_array($permission, app('userPermissions'));
                 });
             }
         }
+    }
+
+    protected function getUser(): ?User
+    {
+        return auth('sanctum')->user() ?? auth('web')->user() ?? null;
     }
 }
