@@ -22,7 +22,6 @@ use Laravel\Octane\Listeners\StopWorkerIfNecessary;
 use Laravel\Octane\Octane;
 
 return [
-
     /*
     |--------------------------------------------------------------------------
     | Octane Server
@@ -36,7 +35,7 @@ return [
     |
     */
 
-    'server' => env('OCTANE_SERVER', 'roadrunner'),
+    'server' => env('OCTANE_SERVER', 'swoole'),
 
     /*
     |--------------------------------------------------------------------------
@@ -132,7 +131,7 @@ return [
     ],
 
     'flush' => [
-        //
+        \Barryvdh\Debugbar\LaravelDebugbar::class,
     ],
 
     /*
