@@ -1,5 +1,5 @@
-require("sweetalert");
-
-import Alpine from 'alpinejs';
-window.Alpine = Alpine;
-Alpine.start();
+import './bootstrap';
+import Alpine from 'alpinejs'
+ 
+window.Alpine = Alpine
+Alpine.start()
