@@ -1,2 +1,2 @@
-<li class="badge">منوی اصلی</li>
+<li class="badge">{{ __('Main menu') }}</li>
 @action('panel.menu_items')
