@@ -46,7 +46,7 @@
                 <label for="permissions">{{__('Role')}}</label>
                 <select name="role_id" id='role_select' class="form-control">
                     <option value="0">{{__('Normal user')}}</option>
-                    @foreach (\App\Models\User\Role::all() as $role)
+                    @foreach (\App\Models\Role::get() as $role)
                         <option value="{{ $role->id }}">{{ $role->title }}</option>
                     @endforeach
                 </select>
